+++ NEWS.md
@@ -1,7 +1,12 @@
-# admiraldev 0.2.1 (UNRELEASED)
-
+# admiraldev 0.3.0
+## New Features
+  - New function `process_set_values_to()` for creating the variables specified
+  by the `set_value_to` argument and catching errors (#70)
+  
 ## Updates of Existing Functions
   - Fix datatable styling for documentation 
+  - The `assert_character_vector()` function gained a `named` argument to check
+  that all elements of the vector are named. (#70)
 
 # admiraldev 0.2.0